logging channel 1 data on sd with time stamp

+++ src/ADC_config.h
@@ -30,7 +30,7 @@ class ADC{
         }
     }
 
-    int Read(){
+    uint16_t Read(){
         analogSetWidth(width);
         return analogRead(pin);
     }
